Added support for models. Added mipmapping. Removing colors.hpp (not in use)

+++ src/graphics/renderer.cpp
@@ -5,7 +5,7 @@
 #include "../utility/debug.h"
 #include "commandPool.h"
 
-Renderer::Renderer(Window& window, std::vector<Vertex>& verticies, std::vector<uint16_t>& indices):
+Renderer::Renderer(Window& window, std::vector<Vertex>& verticies, std::vector<uint32_t>& indices):
 
 vulkanInstance(*window.sdlWindow),
